+++ src/components/periodical-scroll/js/index.js
@@ -24,7 +24,7 @@ export const PERIODICAL_SCROLL = ( ( module, $, window, document ) => {
 	
 	
     module.PERIODICAL_SCROLL               = module.PERIODICAL_SCROLL || {};
-    module.PERIODICAL_SCROLL.version       = '0.0.3';
+    module.PERIODICAL_SCROLL.version       = '0.0.5';
     module.PERIODICAL_SCROLL.documentReady = function( $ ) {
 
 		$( '.uix-periodical-scroll' ).each( function() {
@@ -37,6 +37,11 @@ export const PERIODICAL_SCROLL = ( ( module, $, window, document ) => {
 			const $list       = $this.find( '> ul' );
 			const itemHeight  = $list.find( 'li:first' ).height();
 
+			$this.css({
+				'height': itemHeight + 'px',
+				'overflow': 'hidden'
+				
+			})
 
 			if ( typeof speed === typeof undefined ) {
 				speed = 600;
@@ -46,23 +51,37 @@ export const PERIODICAL_SCROLL = ( ( module, $, window, document ) => {
 				timing = 2000;
 			}	
 			
+			
+			//If there is only one item, add one to complete the seamless loop effect
+			if ( $list.find( 'li' ).length == 1 ) {
+				$list.prepend( $list.find( 'li:first' ).clone() );
+			}
+			
+			
+			//
+			const eachItemAnimOKDelay = 150;
 			const $item = $list.find( '> li' );
 			const moveY = itemHeight*2;
+			
+			
+			
+			//Prevent repetition of animation events
+			TweenMax.killTweensOf( $item );
+			
+			
+			//
 			const tl = new TimelineMax({
-                              onComplete: function() {
-                                  
-                                    TweenMax.set( $item.first(), {
-                                        opacity : 0,
-                                        y       : moveY
-                                    });
-                                  
-                                    setTimeout( function() {
-                                        tl.restart();
-                                    }, timing );
+				              repeat      : -1,
+				              repeatDelay : eachItemAnimOKDelay/1000
 
-                              }
                            });
             
+			
+			//pauses wherever the playhead currently is:
+			tl.pause();
+			setTimeout( function() {
+				tl.play();
+			}, speed );
             
 
 			tl
@@ -71,22 +90,39 @@ export const PERIODICAL_SCROLL = ( ( module, $, window, document ) => {
                 y       : moveY
             },{
                 opacity : 1,
-                y       : 0,
+                y       : 0
             }, timing/1000 ) )
 
             .add( TweenMax.staggerTo( $item, speed/1000, {
                 delay    : timing/1000,
                 opacity  : 0,
                 y        : -moveY,
+				onComplete: function() {
+					TweenMax.set( this.target, {
+						delay    : eachItemAnimOKDelay/1000,
+						opacity : 0,
+						y       : moveY
+					});
+								  
+				}
             }, timing/1000 ), 0 );
             
             
 			
-			$list.on( 'mouseenter', function() { 
+			$item.on( 'mouseenter', function() { 
 				tl.pause();
 			} )
 		    .on( 'mouseleave', function() { 
 				tl.play();
+				
+				if ( $( this ).index() > 0 ) {
+					TweenMax.set( $item.first(), {
+						opacity : 0,
+						y       : moveY
+					});	
+				}
+
+				
 			} );
 			
 
@@ -94,7 +130,8 @@ export const PERIODICAL_SCROLL = ( ( module, $, window, document ) => {
 		});
 		
     };
-
+	
+	
     module.components.documentReady.push( module.PERIODICAL_SCROLL.documentReady );