+++ src/third-party-plugins/esm/Miscellaneous/hashchange.js
@@ -13,6 +13,22 @@
 
 (function($, window, undefined) {
     '$:nomunge';
+	
+	
+  if ( typeof $.browser === typeof undefined ) {
+	jQuery.browser = {};
+	(function () {
+		jQuery.browser.msie = false;
+		jQuery.browser.version = 0;
+		if (navigator.userAgent.match(/MSIE ([0-9]+)\./)) {
+			jQuery.browser.msie = true;
+			jQuery.browser.version = RegExp.$1;
+		}
+	})();  
+  }
+	
+	
+	
     var str_hashchange = 'hashchange',
     doc = document,
     fake_onhashchange, special = $.event.special,